Allow configuration of gerrit.receive in config
There is currently an incompatiblity between Git clients >= 1.8.4.3 and
JGit: https://code.google.com/p/gerrit/issues/detail?id=2296

Summary appears to be recent Git client implementations include an
optimisation around what is packed for a push.  JGit is not ready for
this and results in the push being rejected due to a missing tree
object.

This can be worked around by setting
receive.checkReferencedObjectsAreReachable to false, which was an option
introduced in Gerrit 2.6.  Comments in the JGit source would indicate
these checks are mainly security related, and so can be safely disabled
in a trusted environment (http://goo.gl/BPKlZC).