[TASK] Use Corretto 8 instead of Oracle
Corretto allows to use a OpenJDK 8 and Oracle proves to be an unreliable
download source and problematic due to the new licensing.

As startup takes longer than when using Oracle timeouts have been raised

For some reason curl is no longer included in postgres and h2 builds,
therefore an according general dependency has been added.

The gerrit cookbook requires build-essential which requires mingw which
requires seven_zip (mingw and seven_zip are for windows but well..). The
current seven_zip 3.x requires Chef 13 but we run Chef 12. Pin seven_zip
to <3.0.0 to fix this.